fix: pagina de produtos e seus componentes foram reestruturado afim de adequar-se ao figma
